// src/app/common/enums/sl.enum.ts
export enum SlEnum {
  a = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdhfvdasfwehretefwehrkjsgnhkjyqiueyrughjksd',
  b = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdhfvjkqhgiuyweruqwehrkjsgnhkjyqiueyrughjksd',
  c = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjssdfgrewtbhchgryehrkjsgnhkjyqiueyrughjksd',
  /** save 1 d */
  d = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjwwreyrtydfxhertyruqwehrkjsgnhkjyqiueyrughjksd',
  e = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdnytuyfjndrtyruqwehrkjsgnhkjyqiueyrughjksd',
  /** save 2 f */
  f = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdjtrydtdchdtryrduqwehrkjsgnhkjyqiueyrughjksd',
  g = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakljdrtydfghrtyftdjdtrydfjtdweruqwehrkjsgnhkjyqiueyrughjksd',
  h = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutytrhhgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdhfvjkqhgiuyweruqwehrkjsgnhkjyqiueyrughjksd',
  i = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdhfertyetycfyftycftycykjsgnhkjyqiueyrughjksd',
  /** save 3 j */
  j = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkwueoriueifkldghjakldyrfeioqufrkjsdhfvjkqcytfyctrewywertwetgnhkjyqiueyrughjksd',
  /** save settings history k */
  k = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgm5uj674jdje5uhfjnser5u54ujhfgjhbrthnbrthj65hcytfyctrewywertwetgnhkjyqiueyrughjksd',
  // save 4 l
  l = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutemytjy7i67o78o89qgq34trafzdf34gfsdzff3sgergw4gbrthj65hcytfyctrewywertwetgnhkjyqiueyrughjksd',
  m = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdfksgjlkasdfgernbcxg4retgsdgeryghrjn673565425tresgtxdfg4w6tsegdxgetgnhkjyqiueyrughjksd',
  n = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjgdgmoi4375u9834htijnrjhiyuws90834utjgisjrtklwjk430op8t609grseujtgijwgnhkjyqiueyrughjksd',
  o = 'nblkaueiufngklja9389ureiagjksjfao3i7r9832rhaifaefiug3i7rajhfjweytutegkdfjglhkhbwqkl349468y590oyhikspjtgwoi34uy5t98324utygisoerjgfklwerwertwetgnhkjyqiueyrughjksd',
}